+++ app/src/main/java/com/tosslab/jandi/app/events/entities/ChatCloseEvent.java
@@ -4,13 +4,13 @@ package com.tosslab.jandi.app.events.entities;
  * Created by Steve SeongUg Jung on 15. 4. 9..
  */
 public class ChatCloseEvent {
-    private final int companionId;
+    private final long companionId;
 
-    public ChatCloseEvent(int companionId) {
+    public ChatCloseEvent(long companionId) {
         this.companionId = companionId;
     }
 
-    public int getCompanionId() {
+    public long getCompanionId() {
         return companionId;
     }
 }
