+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqTeam.java
@@ -4,7 +4,7 @@ package com.tosslab.jandi.app.network.models;
  * Created by Steve SeongUg Jung on 14. 12. 11..
  */
 public class ReqTeam {
-    public int teamId;
+    public long teamId;
 
     public ReqTeam(long teamId) {
         this.teamId = teamId;
