+++ app/src/main/java/com/tosslab/jandi/app/network/models/ReqUnshareMessage.java
@@ -4,8 +4,8 @@ package com.tosslab.jandi.app.network.models;
  * Created by justinygchoi on 2014. 7. 28..
  */
 public class ReqUnshareMessage {
-    public int teamId;
-    public int unshareEntity;
+    public long teamId;
+    public long unshareEntity;
 
     public ReqUnshareMessage(long teamId, long unshareEntity) {
         this.teamId = teamId;
